+++ src/torchliter/engine/_types.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 import torch.optim as optim
 
-from .buffer import BufferBase
+from .buffers import BufferBase
 
 __all__ = [
     "MODEL",
